Add description of normalization step
I add a short description of the normalization step done to simplify
the computation of the KL-divergence. I also fixed the variance
scaling to use (1-lamb)**2 instead of (1-lamb). We fixed the bug for
the computation and the results in the paper are with the correct
scaling, but unfortunately the fix did not made it into this repo.